refactor(sloppy): name the turn's steps the way the ecosystem does
Disclosure is a generic collapsible, not an agent concept, so it moves to
components/ui with the rest of the primitives.

The components around it take the names Vercel's AI Elements gives them,
since that is the component library for the SDK this runs on: Reasoning,
Tool and ToolOutput, and Task for the block that holds a turn's steps.
That retires the invented vocabulary, ActivityBlock and WorkPart.

Two names stay put. usage.thoughtSeconds is already persisted, and the
label still reads "Thought for 15s", which is what this kind of product
calls it.
