+++ app/components/sloppy/SloppyMessage.tsx
@@ -1,10 +1,16 @@
 "use client";
 
 import { BookOpen } from "@/components/ui/icon";
-import { messageParts, type AgentMessage } from "@/lib/agent/types";
+import { Disclosure, DisclosureText } from "@/components/ui/disclosure";
+import {
+	messageParts,
+	type AgentContentPart,
+	type AgentMessage,
+} from "@/lib/agent/types";
 import { PanelCard } from "../canvas/panel/PanelCard";
-import { Disclosure, DisclosureText } from "./Disclosure";
-import { ActivityBlock, WorkPart } from "./SloppyActivity";
+import { Reasoning } from "./Reasoning";
+import { Task } from "./Task";
+import { Tool, ToolOutput } from "./Tool";
 import { turnStatus } from "./turnDisplay";
 
 export function UserMessage({ message }: { message: AgentMessage }) {
@@ -22,6 +28,33 @@ function userText(message: AgentMessage): string {
 		.join("\n");
 }
 
+function Step({
+	part,
+	reasoningSeconds,
+	superseded,
+}: {
+	part: AgentContentPart;
+	reasoningSeconds: number | null | undefined;
+	superseded: boolean;
+}) {
+	switch (part.type) {
+		case "reasoning":
+			return (
+				<Reasoning
+					text={part.text}
+					seconds={reasoningSeconds}
+					superseded={superseded}
+				/>
+			);
+		case "tool-call":
+			return <Tool toolName={part.toolName} input={part.input} />;
+		case "tool-result":
+			return <ToolOutput output={part.output} />;
+		default:
+			return null;
+	}
+}
+
 function ReplyText({ text }: { text: string }) {
 	return (
 		<p className="whitespace-pre-wrap break-words text-label text-foreground">
@@ -31,8 +64,8 @@ function ReplyText({ text }: { text: string }) {
 }
 
 /**
- * How the turn went sits in one collapsible block; what Sloppy said sits in the
- * card below it, so the reply survives the block closing.
+ * How the turn went sits in one collapsible task; what Sloppy said sits in the
+ * card below it, so the reply survives the task closing.
  */
 export function AgentTurn({
 	messages,
@@ -52,14 +85,14 @@ export function AgentTurn({
 	idPrefix: string;
 }) {
 	const parts = messages.flatMap(messageParts);
-	const work = parts.filter((part) => part.type !== "text");
+	const steps = parts.filter((part) => part.type !== "text");
 	const said = parts.filter((part) => part.type === "text");
 	const newest = parts.at(-1);
 
 	return (
 		<>
-			{(streaming || systemPrompt || work.length > 0) && (
-				<ActivityBlock
+			{(streaming || systemPrompt || steps.length > 0) && (
+				<Task
 					streaming={streaming}
 					status={turnStatus(parts)}
 					seconds={workSeconds}
@@ -74,15 +107,15 @@ export function AgentTurn({
 							<DisclosureText>{systemPrompt}</DisclosureText>
 						</Disclosure>
 					)}
-					{work.map((part, index) => (
-						<WorkPart
+					{steps.map((part, index) => (
+						<Step
 							key={`${idPrefix}-${index}`}
 							part={part}
-							thoughtSeconds={thoughtSeconds}
+							reasoningSeconds={thoughtSeconds}
 							superseded={part !== newest}
 						/>
 					))}
-				</ActivityBlock>
+				</Task>
 			)}
 			{said.length > 0 && (
 				<PanelCard>
